fix warning C4777: '_snprintf' : format string '%lX' requires an argument of type 'unsigned long', but variadic argument 1 has type 'XMP_OptionBits'

+++ xmpsdk/src/XMPMeta.cpp
@@ -238,7 +238,7 @@ DumpNodeOptions	( XMP_OptionBits	 options,
 	} else {
 	
 		OutProcNChars ( "(0x", 3 );
-		OutProcHexInt ( options );
+		OutProcHexInt ( (unsigned long)options );
 		OutProcNChars ( " :", 2 );
 
 		XMP_OptionBits mask = 0x80000000;
